Refactored migration to separate migration status
