+++ src/databricks/labs/ucx/source_code/lsp.py
@@ -13,9 +13,9 @@ from databricks.labs.blueprint.logger import install_logger
 from databricks.sdk.service.workspace import Language
 
 from databricks.labs.ucx.hive_metastore.table_migrate import (
-    MigrationIndex,
     MigrationStatus,
 )
+from databricks.labs.ucx.hive_metastore.migration_status import MigrationIndex
 from databricks.labs.ucx.source_code.base import (
     Advice,
     Advisory,
