+++ tests/unit/source_code/test_languages.py
@@ -1,7 +1,7 @@
 import pytest
 from databricks.sdk.service.workspace import Language
 
-from databricks.labs.ucx.hive_metastore.table_migrate import MigrationIndex
+from databricks.labs.ucx.hive_metastore.migration_status import MigrationIndex
 from databricks.labs.ucx.source_code.base import Fixer, Linter
 from databricks.labs.ucx.source_code.languages import Languages
